requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
firebase-admin
python-dotenv
